+++ lib/services.ts
@@ -65,7 +65,7 @@ export class Services {
   /**
    * List services applied to a domain.
    *
-   * This API is paginated. Call `applyService.iterateAll(account, domain, params)` to get an asynchronous iterator over individual items across all pages. You can also use `await applyService.collectAll(account, domain, params)` to quickly retrieve all items across all pages into an array. We suggest using `iterateAll` when possible, as `collectAll` will make all requests at once, which may increase latency and trigger rate limits.
+   * This API is paginated. Call `appliedServices.iterateAll(account, domain, params)` to get an asynchronous iterator over individual items across all pages. You can also use `await appliedServices.collectAll(account, domain, params)` to quickly retrieve all items across all pages into an array. We suggest using `iterateAll` when possible, as `collectAll` will make all requests at once, which may increase latency and trigger rate limits.
    *
    * GET /{account}/domains/{domain}/services
    *
@@ -75,7 +75,7 @@ export class Services {
    * @param domain The domain name or id
    * @param params Query parameters
    */
-  applyService = (() => {
+  appliedServices = (() => {
     const method = (
       account: number,
       domain: string,
@@ -119,7 +119,7 @@ export class Services {
    * @param service The service sid or id
    * @param params Query parameters
    */
-  appliedServices = (() => {
+  applyService = (() => {
     const method = (
       account: number,
       domain: string,