+++ CHANGELOG.md
@@ -191,8 +191,9 @@ literal string as a constant string type.
 
 This is a major change that brings TypeScript support as well as many
 quality-of-life improvements and internal improvements. For details on important
-changes and how to migrate, see [UPGRADE.md](./UPGRADE.md). For the full
-breakdown of all changes in detail, see the PR (#170).
+changes and how to migrate, see
+[UPGRADE.md](https://github.com/dnsimple/dnsimple-node/blob/v7.0.0/UPGRADE.md).
+For the full breakdown of all changes in detail, see the PR (#170).
 
 ### Changed