+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is
 based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 14.0.0 - 2026-09-22
 
 ### Added